git-svn-id: https://svn.apache.org/repos/asf/poi/tags/REL_3_5_BETA6@782402 13f79535-47bb-0310-9956-ffa450edef68

+++ src/testcases/org/apache/poi/hssf/record/TestCommonObjectDataSubRecord.java
@@ -56,7 +56,7 @@ public final class TestCommonObjectDataSubRecord extends TestCase {
 		CommonObjectDataSubRecord record = new CommonObjectDataSubRecord();
 
 		record.setObjectType(CommonObjectDataSubRecord.OBJECT_TYPE_LIST_BOX);
-		record.setObjectId((short) 1);
+		record.setObjectId( 1);
 		record.setOption((short) 1);
 		record.setLocked(true);
 		record.setPrintable(false);
